requirements.txt
pyautogen>=0.7.6
openai>=1.3.7
requests